move procedural filter processing from renderer to browser
